requirements.txt
scenedetect
opencv-python<4.12
numpy>=1.26.4,<2.0
transformers
Pillow
torch
chromadb